requirements.txt
aadict==0.2.3
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
asset==0.6.13
async-timeout==4.0.3
attrs==23.1.0
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
# dataclasses==0.8
datasets==2.14.5
dill==0.3.7
# dm-tree==0.1.7
docker-pycreds==0.4.0
einops==0.7.0
# fastmoe==1.1.0
# fastmoe==1.1.0
filelock==3.9.0
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.11
GitPython==3.1.42
globre==0.1.5
gmpy2==2.1.2
huggingface-hub==0.18.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.1
# mkl-fft==1.3.8
# mkl-random==1.2.4
# mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.2
multiprocess==0.70.15
networkx==3.1
numexpr==2.8.7
numpy==1.26.0
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.0.3
pathtools==0.1.2
Pillow==10.0.1
pip==23.2.1
protobuf==3.20.3
psutil==5.9.0
pyarrow==11.0.0
pycparser==2.21
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0
regex==2023.10.3
requests==2.31.0
sacremoses==0.0.53
safetensors==0.3.2
scikit-learn==1.2.2
scipy==1.11.3
sentry-sdk==1.40.4
setproctitle==1.2.2
setuptools==68.0.0
six==1.16.0
smmap==5.0.0
# speedrun==0.1
sympy==1.11.1
threadpoolctl==3.3.0
tokenizers==0.13.3
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.31.0
triton==2.1.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wandb==0.16.3
wheel==0.41.2
xxhash==2.0.2
yarl==1.8.1
zipp==3.17.0
